test: replace test written in JS with TS
